+++ CMakeLists.txt
@@ -135,7 +135,9 @@ configure_file(src/constants.cpp ${RTT_CONSTANTS_SOURCE_FILE} COPYONLY)
 ###########################
 
 add_library(roboteam_utils
+        src/LineBase.cpp
         src/Line.cpp
+        src/LineSegment.cpp
         src/Vector2.cpp
         src/Angle.cpp
         src/Position.cpp
